fix: aarch64 EA_View enemyorigin pointer truncation in BotAimAtEnemy
Line 20366 had `EA_View(v21, (int)bs->enemyorigin)` — the (int) cast
truncates the bs->enemyorigin float[3] pointer to 32 bits before the
function call's pointer parameter conversion sign-extends garbage.
Crashed CTF mode at AINode_Battle_NBG → BotAimAtEnemy → EA_View+0x38.

Removed the cast so the float* decay reaches EA_View correctly.

Source-commit: d8a0aeca05a8036a7a40783d31e7e121ae34301a

+++ botlib/botlib.c
@@ -20363,7 +20363,7 @@ void BotAimAtEnemy(bot_state_t *bs)
       v21 = bs->client;
       *(int *)&bs->enemyorigin[0] = v20;
       *(int *)&bs->enemyorigin[2] = bs->_i4244;
-      EA_View(v21, (int)bs->enemyorigin);
+      EA_View(v21, bs->enemyorigin);
     }
   }
 }
